requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-filter
pillow
dj-database-url
psycopg[binary]
